+++ templates/core/terraform/main.tf
@@ -146,6 +146,7 @@ module "resource_processor_vmss_porter" {
   key_vault_id                                     = azurerm_key_vault.kv.id
   subscription_id                                  = var.arm_subscription_id
   resource_processor_number_processes_per_instance = var.resource_processor_number_processes_per_instance
+  resource_processor_vmss_sku                      = var.resource_processor_vmss_sku
 
   depends_on = [
     module.azure_monitor,